// wren-ui/src/apollo/client/graphql/__types__.ts
export type Scalars = {
  ID: string;
  String: string;
  Boolean: boolean;
  Int: number;
  Float: number;
  DialectSQL: any;
  JSON: any;
};
// ...
export type Instruction = {
  __typename?: 'Instruction';
  createdAt: Scalars['String'];
  id: Scalars['Int'];
  instruction: Scalars['String'];
  isDefault: Scalars['Boolean'];
  projectId: Scalars['Int'];
  questions: Array<Scalars['String']>;
  updatedAt: Scalars['String'];
};